refactor: extract user types to types/user.ts
