+++ frontend/src/services/profile.service.ts
@@ -1,35 +1,7 @@
 import { ResultOf } from "@/lib/result";
+import type { UserProfile, UserPost, UserSettings, UpdateProfileRequest } from "@/types/user";
 
-export type UserProfile = {
-  id: string;
-  userName: string;
-  email: string;
-  firstName: string;
-  lastName?: string | null;
-  birthdate?: string | null;
-  roleName?: string | null;
-};
-
-export type UserPost = {
-  id: string;
-  title: string;
-  content: string;
-  status: number;
-  authorId: string | null;
-  createdAt: string;
-};
-
-export type UserSettings = {
-  id: string;
-  birthdatePrivacy: "public" | "private";
-};
-
-export type UpdateProfileRequest = {
-  firstName?: string;
-  lastName?: string;
-  userName?: string;
-  birthdate?: string;
-};
+export type { UserProfile, UserPost, UserSettings, UpdateProfileRequest };
 
 class ProfileService {
   private readonly baseUrl = "/api";
